Implement better object based system for triggers
After realizing that a system based on parsing strings and etc wasn't
sustainable (through trying to get the menu to work as a trigger), I
switched over to an object based system where you just put the data
directly in the object in the file and it just works in the program.
Still not finished redoing everything, and I still need to work on
branching.